+++ x/oracle/keeper/ballot.go
@@ -1,8 +1,6 @@
 package keeper
 
 import (
-	"sort"
-
 	"github.com/NibiruChain/nibiru/collections"
 
 	"github.com/NibiruChain/nibiru/x/oracle/types"
@@ -10,16 +8,23 @@ import (
 	sdk "github.com/cosmos/cosmos-sdk/types"
 )
 
-// OrganizeBallotByPair collects all oracle votes for the period, categorized by the votes' pair parameter
-func (k Keeper) OrganizeBallotByPair(ctx sdk.Context, validatorsPerformance map[string]types.ValidatorPerformance) (ballots map[string]types.ExchangeRateBallot) {
+// mapBallotByPair collects all oracle votes for the period, categorized by the votes' pair parameter
+// and removes any votes that are not part of the validator performance map.
+//
+// NOTE: **Make abstain votes to have zero vote power**
+func (k Keeper) mapBallotByPair(
+	ctx sdk.Context,
+	validatorsPerformance map[string]types.ValidatorPerformance,
+) (ballots map[string]types.ExchangeRateBallot) {
 	ballots = map[string]types.ExchangeRateBallot{}
 
-	// Organize aggregate votes
-	aggregateHandler := func(voterAddr sdk.ValAddress, vote types.AggregateExchangeRateVote) (stop bool) {
+	for _, value := range k.Votes.Iterate(ctx, collections.Range[sdk.ValAddress]{}).KeyValues() {
+		voterAddr, vote := value.Key, value.Value
+
 		// organize ballot only for the active validators
-		if claim, ok := validatorsPerformance[vote.Voter]; ok {
+		if validatorPerformance, ok := validatorsPerformance[vote.Voter]; ok {
 			for _, tuple := range vote.ExchangeRateTuples {
-				power := claim.Power
+				power := validatorPerformance.Power
 				if !tuple.ExchangeRate.IsPositive() {
 					// Make the power of abstain vote zero
 					power = 0
@@ -35,25 +40,13 @@ func (k Keeper) OrganizeBallotByPair(ctx sdk.Context, validatorsPerformance map[
 				)
 			}
 		}
-
-		return false
-	}
-
-	for _, vote := range k.Votes.Iterate(ctx, collections.Range[sdk.ValAddress]{}).KeyValues() {
-		aggregateHandler(vote.Key, vote.Value)
-	}
-
-	// sort created ballot
-	for pair, ballot := range ballots {
-		sort.Sort(ballot)
-		ballots[pair] = ballot
 	}
 
 	return
 }
 
-// ClearBallots clears all tallied prevotes and votes from the store
-func (k Keeper) ClearBallots(ctx sdk.Context, votePeriod uint64) {
+// clearBallots clears all tallied prevotes and votes from the store
+func (k Keeper) clearBallots(ctx sdk.Context, votePeriod uint64) {
 	// Clear all aggregate prevotes
 	for _, prevote := range k.Prevotes.Iterate(ctx, collections.Range[sdk.ValAddress]{}).KeyValues() {
 		if ctx.BlockHeight() > int64(prevote.Value.SubmitBlock+votePeriod) {
@@ -73,18 +66,16 @@ func (k Keeper) ClearBallots(ctx sdk.Context, votePeriod uint64) {
 	}
 }
 
-// ApplyWhitelist updates the whitelist by detecting possible changes between
+// applyWhitelist updates the whitelist by detecting possible changes between
 // the current vote targets and the current updated whitelist.
-func (k Keeper) ApplyWhitelist(ctx sdk.Context, whitelist []string, voteTargets map[string]struct{}) {
-	// check is there any update in whitelist params
+func (k Keeper) applyWhitelist(ctx sdk.Context, whitelist []string, whitelistedPairsMap map[string]struct{}) {
 	updateRequired := false
-	// fast path
-	if len(voteTargets) != len(whitelist) {
+
+	if len(whitelistedPairsMap) != len(whitelist) {
 		updateRequired = true
-		// slow path, we need to check differences
 	} else {
 		for _, pair := range whitelist {
-			_, exists := voteTargets[pair]
+			_, exists := whitelistedPairsMap[pair]
 			if !exists {
 				updateRequired = true
 				break