+++ x/oracle/keeper/msg_server.go
@@ -102,7 +102,7 @@ func (ms msgServer) AggregateExchangeRateVote(goCtx context.Context, msg *types.
 
 	// check all pairs are in the vote target
 	for _, tuple := range exchangeRateTuples {
-		if !ms.IsVoteTarget(ctx, tuple.Pair) {
+		if !ms.IsWhitelistedPair(ctx, tuple.Pair) {
 			return nil, sdkerrors.Wrap(types.ErrUnknownPair, tuple.Pair)
 		}
 	}