+++ x/oracle/keeper/update_exchange_rates_test.go
@@ -212,7 +212,7 @@ func TestOracleTally(t *testing.T) {
 		}
 	}
 
-	tallyMedian := keeper.Tally(input.Ctx, ballot, input.OracleKeeper.RewardBand(input.Ctx), validatorClaimMap)
+	tallyMedian := keeper.Tally(ballot, input.OracleKeeper.RewardBand(input.Ctx), validatorClaimMap)
 
 	require.Equal(t, validatorClaimMap, expectedValidatorClaimMap)
 	require.Equal(t, tallyMedian.MulInt64(100).TruncateInt(), weightedMedian.MulInt64(100).TruncateInt())
@@ -486,7 +486,6 @@ func TestVoteTargets(t *testing.T) {
 	params.Whitelist = []string{common.Pair_NIBI_NUSD.String(), common.Pair_BTC_NUSD.String()}
 	input.OracleKeeper.SetParams(input.Ctx, params)
 
-	// clear tobin tax to reset vote targets
 	for _, p := range input.OracleKeeper.Pairs.Iterate(input.Ctx, collections.Range[string]{}).Keys() {
 		input.OracleKeeper.Pairs.Delete(input.Ctx, p)
 	}
@@ -505,7 +504,7 @@ func TestVoteTargets(t *testing.T) {
 	require.Equal(t, uint64(0), input.OracleKeeper.MissCounters.GetOr(input.Ctx, keeper.ValAddrs[2], 0))
 
 	// vote targets are {govstable, btcstable}
-	require.Equal(t, []string{common.Pair_BTC_NUSD.String(), common.Pair_NIBI_NUSD.String()}, input.OracleKeeper.GetVoteTargets(input.Ctx))
+	require.Equal(t, []string{common.Pair_BTC_NUSD.String(), common.Pair_NIBI_NUSD.String()}, input.OracleKeeper.GetWhitelistedPairs(input.Ctx))
 
 	// delete btcstable
 	params.Whitelist = []string{common.Pair_NIBI_NUSD.String()}
@@ -523,7 +522,7 @@ func TestVoteTargets(t *testing.T) {
 	require.Equal(t, uint64(1), input.OracleKeeper.MissCounters.GetOr(input.Ctx, keeper.ValAddrs[2], 0))
 
 	// btcstable must be deleted
-	require.Equal(t, []string{common.Pair_NIBI_NUSD.String()}, input.OracleKeeper.GetVoteTargets(input.Ctx))
+	require.Equal(t, []string{common.Pair_NIBI_NUSD.String()}, input.OracleKeeper.GetWhitelistedPairs(input.Ctx))
 
 	exists := input.OracleKeeper.Pairs.Has(input.Ctx, common.Pair_BTC_NUSD.String())
 	require.False(t, exists)